Add service settings page for data retention
This page is intended for users to self-serve data retention for their
service, although while we are building this out we're restricting it to
platform admins still.

Users will be able to set the data retention for their service's emails,
text messages and letters (via a single value).

We will retain the platform admin custom data retention page as this can
set per-channel retentions, which is complexity we don't think users
need at this stage.