Update error message for length constraints
We are updating the error message for >255 characters entered for a
security key name. This uses the `Length` validator from WTForms. We
have a pattern for this error message, so rather than hard-coding the
exact message let's create a validator which applies the pattern
consistently.

Let's save some future work by applying this across the board now where
we can safely do so. Where the error messages might be more complex, we
leave the old hard-coded message in place and will tackle it in the
future when we do a page that does email validation.