Disable form for high volume services
We only want services that are going to send <1mil notifications a year
to be able to change their data retention in a self-service way. High
volume services will need to contact us to change their data retention.

We determine a high volume service in a few ways:

1) If they've told us they will send over 1 million notifications in a
year
2) If they sent over 1 million notifications last financial year
3) If we expect them to send over 1 million notifications this year,
based on their current usage projected to the end of the financial year.