Make font size smaller when numbers are very big
We have a problem where big numbers overflow the width of their columns.
This happens when the numbers are bigger than we anticipated (for
example when we designed the page we never imagined anyone would spend
tens of millions of £ on text messages in a single year).

This happens in 3 places:
- at the bottom of the service dashboard, which links to…
- the usage page for a service
- the usage page for an organisation

This commit adds some logic to reduce the font size for all the totals
if any of them get so big that they overflow the width of the column.
This gives us enouigh space to display the biggest numbers we currently
store in the platform.

Because we use tabular numbers for these elements we don’t need to worry
about the width of 1,111,111 versus 9,999,999.

A function which takes 2 arguments, `smaller` and `smallest` isn’t a
great piece of API design, but if we want to refactor that it should be
as a separate PR. This commit just works within the confines of the
existing argument names.

Similarly we can’t wrap this logic up in the `big_number` component
because it doesn’t know about:
- the width of the column in which it’s shown
- whether other numbers on the page are too big