Strip error message links in error summary
In our error messages for invalid gov emails, we link to our
documentation around who can use Notify. But we need to not render the a
tag inside the error summary, because each entry is an anchor to the
appropriate error input.

So we strip out tags inside error messages when they're rendered in the
error summary.