Disable custom retention for trial mode services
If a service is in trial mode, we don't have a good idea yet about how
many messages they're planning to send each year. As part of the go-live
price they will give us an estimated usage, which we can then use to
decide whether they can set their own retention or not.