Always treat UUIDs in URLs as lowercase
In 554a852e2d2d1e5c0315af6ea1feaf179e5d42d2 we overrode `to_python`
method of Flask’s UUID URL convertor to return strings, not `UUID`
instances.

This is because it’s easier for us to deal with strings in our test
assertions.

However one advantage of dealing with `UUID` instances is that when they
are cast to `str` they are always lowercase. This is behaviour that we
didn’t match in our overridden method.

Flask’s UUID URL convertor matches on lowercase, uppercase or mixed
case UUIDs (which is technically correct because they should be treated
as hexadecimal values).

This means that when we passed UUIDs to other parts of the codebase it
was possible for them to be strings with uppercase characters in.

This caused a particular problem with our cache keys, which were
expecting consistent casing.

While we are fixing this in the caching code (see https://github.com/alphagov/notifications-utils/pull/1064/files)
I think it’s also a good idea to still be consistent with Flask’s
default UUID URL convertor, to prevent unpredictable bugs like this in
the future.