Set gunicorn keepalive at 90 seconds
We haven't been setting this previously so we are currently
on the default of 2 seconds:
https://docs.gunicorn.org/en/stable/settings.html#keepalive

We are setting this value to be more than 60 seconds because 60
seconds is the default idle timeout of AWS ALBs. We have been
seeing some 502s in our preview ECS environment that we believe
is being caused by a lower keepalive time than the ALBs idle
timeout. The evidence for this is from
https://repost.aws/knowledge-center/elb-alb-troubleshoot-502-errors
https://docs.aws.amazon.com/elasticloadbalancing/latest/application/load-balancer-troubleshooting.html#http-502-issues
https://repost.aws/questions/QUqP4BHC9iQ0uIB69M7QrjSg/502-errors-with-application-load-balancer-idle-timeout-apache2-keep-alive-timeout

The next question is, why isn't this happening on the PaaS? The reason
appears to be that the PaaS gorouter has disabled keepalives:
https://github.com/alphagov/paas-cf/blob/83961d94ee9072e8899b1ae55dad1981ed35c94c/manifests/cf-manifest/operations.d/310-router.yml#L33

This should also mean that this change on the PaaS won't have
any affect because keepalives aren't being used for the connection
between the gorouter and our apps anyway.

If this change fixes our 502s, then we will need to make an equivalent
change to our other apps