Fix missing `url_for`
This was redirecting to `/user-profile/email/main.user_profile_email`
(the name of the endpoint in Python) rather than `/user-profile/email`
(the URL specified in the Flask decorator) because the call to `url_for`
was missing.

There was also no test for this so that’s why it slipped through.

A good example of why breaking up changes into smaller commits and pull
requests is a nice thing 😅
https://github.com/alphagov/notifications-admin/pull/117/files