Focus first times radio when moving from days
Causes legend of fieldset to be announced so this
also moves the help text into the fieldset and
links it through aria-describedby.

Initially tried to focus the times view but this
caused all its contents to be announced with some
screen readers.

Also removes code that prevented scrolling when
times view was focused as no longer needed.