Paginate template history
We know some services have huge numbers of template versions, and this
page effectively breaks for those users.

Let's put pagination on this to reduce the HTML size for those services,
and keep the page a bit more responsive.

Note that this is a quick win rather than a full fix, as the API doesn't
support pagination on this endpoint and so still sends the entire
template history across. We paginate on the admin side only.

We should add pagination to the API to fix this properly, but it's more
complex since we're working across two apps and have to tweak our
template version caching as well.