Fix JSDOM's lack of support for form submission
JSDOM doesn't includes support for form
submission, or any type of navigation by design.

Because of this we have to use a hack to check for
form submissions, instead using a technique from
the preventDuplicateFormSubmissions test where we
instead track the error JSDOM emits for the lack
of that implementation.