Update text for link to days view
'Back to days' is used to piggyback on the
'back-link' pattern we, and lots of other
government services, use.

After some discussion, we feel that using it out
of its normal context could add as much confusion
as it solves. This removes the use of the 'Back
to' prefix and makes it more of a
'call-to-action'.