Support enabling sentry tracing through header
At the moment Sentry is configured to trace 1% of requests, which it
does effectively at random.

But when we hit some requests that don't perform well, it can be useful
to have a way to force Sentry to trace it so that we can reliably
investigate those performance issues.

This patch lets us add a query parameter to the URL to force sentry to
trace the request (which will also pass through to the API as well), so
that we can quickly investigate potential issues. Alternatively, we can
pass an 'X-Notify-Sentry-Trace' header which will help trace multiple
requests more easily.