Rename from_database_object to get_preview_for_templated_letter
Also remove from_utils_template - it was not doing much, just
calling get_preview_for_templated_letter with placeholder values
filled in.

So instead of two TemplatePreview class methods we now have one:
 get_preview_for_templated_letter

I renamed that method so that it says what it does.

In next commits I will remove, and if needed, move around,
other methods in that class.