Avoid `werkzeug.request`
This key doesn't exist when we run the app using gunicorn/gevent (but is
available on the local flask dev server).

Let's prefer to use standard keys; `wsgi_environ` always contains the
request headers in a standard format as native keys in that dict. Let's
query that instead for our custom sentry trace header.