Remove scheduled recipients when real ones show
When a job starts sending we update various parts of the page, for
example to replace the ‘Cancel’ button with the list of notifications.

We do this by loading new HTML from the server and replacing the HTML
on the page.

We can’t do this for the recipients of a scheduled job. This is because
the recipients are in a table which is enhanced with Javascript. When
time we poll for new content from the server we replace the enhanced
table with a non-enhanced one.

Just can’t just re-initialise the table enhancements after polling,
because stuff like the user’s horizontal scroll position would be lost.

So instead this commit adds another little bit of Javascript which
removes an element (the scheduled recipients) from the page when it sees
the presence of another element (the actual list of notifications).