Look at any parent for width of scrollable table
We determine the width of the scrollable table by looking at the width
of the element that encloses it. It doesn’t do this automatically
because it is absolutely-positioned.

Previously we were looking for:
- the `<main>` element
- where it was the immediate parent of the table

Now that we have wrapped the table in another `<div>` this doesn’t work,
because the `<main>` tag is no longer the immediate parent.

Instead let’s make this less strict and look at the immediate parent, no
matter what element it is.