Email org user about continuing journey
If the org approver isn't sure that the service is unique, or doesn't
think the service name is easy to understand, then they need to take the
journey offline and talk to the service user directly.

To help them pick the approval up again, we send an email to the org
member with this information and links back to the 'make service live'
flow.