Email user about rejected go-live request
If the org member rejects the service's request to go live, send an
email to the service user that submitted the request.

We will shortly include a rejection reason that will be collected from
the org member. The email will include contact details for the org
member so that the service member can get in touch if needed.