Remove old route for adding letter templates
Adding a letter template is different because we first create an
‘untitled’ template which the user can go and edit.

For emails and text messages we create the template from the user’s
input, including the name and content.

This means no-one should be using the ‘add template’ view to create
letter templates.

By enforcing this we can make the code in the add template view a bit
simpler since it no longer has to consider letters.