Remove `extract_cloudfoundry_config`
We no longer need to provide the REDIS_URL from VCAP_SERVICES as
it will always come directly from the `REDIS_URL` environment variable.
Because this was the only use for the `extract_cloudfoundry_config`
function, we can delete the entire function!

I then couldn't figure out any value in having the test_config.py
tests. They appear to only test a particular  one of our config values can
be set based on the value of an environment variable. I wasn't sure
this was worth keeping or not.