Remove redundant redis manifest code
We pass in a REDIS_URL for all PaaS environments so we don't need
to check if it is defined

Our check for redis being enabled was a temporary thing for
our migration to our new redis. Going forward, we want redis to
always be enabled in the PaaS and aren't planning on changing
that so don't need to redefine that in the manifest (redis is
enabled by default in the config.py file).