validate service name
ensure that it is able to be converted into a valid email. In some
cases, the service name won't be used (if the service has a custom email
sender set up) but I think this is still a reasonable validation as
covering all the state changes to ensure that services are valid if
they're used for a from email would be susceptible to bugs.

also validate the length of the email address generated stays within the
email spec (320 characters including quoted display name.

we can't do this in the `validate_email_address` function since that has no
concept of the display name part - so we need to build it up in the
admin app where the form field has the user entered part and the
normalised part.