fix copy letter template page
we removed the ability to call "add template" for letters in fe7e9a32

however, this inadvertently broke the copy template page, which calls
through to add template. change the guard to refer to the endpoint
itself to allow copying.

Notably in the commit mentioned above we removed some complexity around
validating QR code lengths. This can still be removed since the previous
template you're copying will have already passed those checks when
edited.

we also need to add the TemplateNameMixin back into the
LetterTemplateForm since we need to copy across the name when creating a
new copied template. We don't use this form when creating a new
template, and when editing an existing template we pre-populate the form
with the prev contents of the template, so the name simply gets passed
through.