pass through the name on the edit letter template form
we don't let you rename letter templates from this edit page. However,
when you copy a template we show you this page with the data prefilled
to match that of the existing template you're copying. When you save we
create the new template so we need your copy name, but lets keep it
hidden to keep this page consistent with if you edit the tepmlate
normally.