Save original PDF for letter attachments
We had code to save the sanitised letter attachment and the original
letter attachment (before we have changed it) but we were saving the
sanitised version in both cases. This fixes the code to ensure we are
saving the original file as well as the sanitised version.

Also removes a few unused test fixtures.