Make `letter_languages` an expected key
This key should always be set on templates now, so let's stop using a
safe `get`. This gave us a broken test, as `template_json` was not
providing values for `letter_languages` or either of the welsh keys. Our
tests were no longer representative.

This fixes `template_json` and updates the code to be less forgiving,
since all of our templates now have a `letter_languages` value.