Fix link between error summary and file upload
When we generate the links in the error summary,
we set them to the id of the field with that error
so clicking them jumps you to it.

Replacing the upload field with a button broke
this link because the field linked to was now
hidden. This copies the field id to the button so
any error links will still work.