Bump Node to 16.16.0 LTS
This is the current 16 LTS version. It also, with PR 1791 on
notifications-aws, brings both versions back in sync and adds comments
to help keep them that way.

With them both in sync we should have faster and more reliable test builds.

If bumping the node version in `.nvmrc`, then you also must update:

* `docker/Dockerfile` in this repo
* ``concourse/docker/unit-tests-image-node16.Dockerfile` in notifications-aws