Fix images in IE11 & mobile viewports
IE11 didn't respect the use of `max-width: 100%`
and instead went by the widths set on each of the
images. This changes to `width: 100%` to make sure
they always fill the space in their container
(which I think was always the intent of
`max-width` anyway) but in a way that IE11
respects over the width on the element.

We also didn't do much work on this component on
smaller screens and the current implementation
makes the images small enough that some users
might not be able to see the text in them
properly. This stacks them vertically which gives
enough space for their text to be legible.