Show email sender consistently for both options
This commit puts the example of the thing you are selecting (or is
currently selected) next to the respective radio buttons.

This means the content on the page has to do less work.

To make the this work for the ‘custom’ option, I have re-used the status
component we use for updating the text message character count. This
means it gives a live preview of the sender name and email address
changing as you type.