Validate very generic sender names
We want recipients to know who the emails we send are coming from. In my
experience this isn’t always obvious when receiving emails from
companies.

We can help mitigate this with good content on the page. But we can also
validate out some common ‘bad’ sender names (based on a quick skim of
my inbox).