Update welsh content when breaking changes
When a breaking change is made to a template, we display an interstitial
page asking the user to confirm the change (new placeholders added).

This page contains some hidden form fields that contain the updated
subject/content. These fields always held the English content, which
would end up overwriting the Welsh content.

Change the form fields to read the subject/content from the submitted
form, which will be the correct English or Welsh data depending on which
language is being edited.