Lift and shift utils.LetterImagePreview
This subclass is only used by admin. We use this class to display PNGs
on the page that represent a preview of a blank letter. On top of those
PNGs we overlay various buttons to edit different parts of the letter,
eg the postage, from address, content etc.

Where those buttons are, is very much a design decision of the admin
app.

As we're looking to add another 'Edit' button for the English pages of
the letter, it becomes tricky to identify where that button should be
displayed as we can't just use a static offset from the top of the first
page. The English pages of a bilingual letter could be page 2, 3, 4, 5
etc. So we need to be able to dynamically insert the button.

Rather than updating utils and passing some data/indicator between admin
and utils, let's just have admin own this subclass entirely.