Move letter template edit buttons
When editing a letter template, we display a number of UI elements
overlaid on the template so that they appear close to the related
content they edit.

For bilingual letters we need to display an 'Edit' button on the first
page of both Welsh and English pages. The first English page could be on
any arbitrary page number, and so a static CSS position offset from the
top of the template (as we currently do for the first edit button) isn't
feasible.

This change brings the UI overlay into the same (jinja) template that
renders the letter preview. This lets us insert the English language
edit button on the first English page.