Remove `broadcast` as a template type
We are removing emergency alerts since it’s been split into its own
codebase.

The first step is to remove `broadcast` as a template type (‘broadcast’,
as in ‘cell broadcast’ was the initial name for emergency alerts).

Removing it as a template type means that users will no longer be able
to see or create message templates for emergency alerts.