Remove `broadcast` as a service permission
We are removing emergency alerts since it’s been split into its own
codebase.

This commit removes `broadcast` from the list of service permissions in
this app.

This also means removing anything which tries to check for this
permission.

This catches the rest of the view-level code which was not part of the
settings page or the pages for viewing/sending alerts.