Remove service properties
We are removing emergency alerts since it’s been split into its own
codebase.

These properties were only used when changing the settings for emergency
alert services, which is no longer possible since we removed the views
to do so.

The API will still be returning this data, but removing the properties
here means that we will ignore it.