Estimate cost of text message on template page
At the moment only people who are editing templates get to see the
estimated cost (added in https://github.com/alphagov/notifications-admin/pull/3763/files)

We think this is causing people to get confused about why they are being
charged for sending more than one message.

We think that by also showing this to people who are just looking at the
template we can help people understand the cost better.

For now not showing this to people who are only sending messages, since
we have a long-standing principle that they shouldn’t be considering
the cost when deciding whether or not to send a message (either the
message needs sending or it doesn’t).