Simplify 'copy template' journey
The current 'copy template' journey is a bit more of a 'pre-filled
create new template' journey. The user is shown a list of all templates
they can see, and when they select one, we take them to the 'add new
template' page with the relevant form fields pre-filled.

This means that all data from the source template needs to be shown on
the page in order to get copied into the new template. As templates get
more complex (eg letter templates with welsh and english pages, and
possibly attachments), this falls outside of our existing design
pattern.

Instead we can simply show confirmation of the template they've
selected, by showing the template preview, and give them a chance to
provide a new name for the template copy (but pre-filled with a sensible
copy name).

This means that we can make sure all of the data is copied correctly
behind-the-scenes. When a user confirms that they're copying the correct
template, we still do a POST to our API to create a new template, but
populated with all of the data from the source data automatically.

To minimise the changeset, I am leaving letter attachments aside for now
still. That will be handled in a future PR.