Allow multiple distinct services to be mocked
`mock_get_service` uses a naive implementation which returns a fixed set
of service JSON, with a service_id depending on what is passed into the
`get_service` call.

This makes it difficult to test code that touches multiple services with
different attributes.

This updates `mock_get_service` to use a dictionary underneath that
can store custom service JSON blobs. If a request is made for a service
that isn't in the explicit dict, we return a default service still.