Add guard against services with no organisation
Code elsewhere was assuming that the service has an organisation by the
time it gets to this point.

So when this method returns `True` early for platform admin users it’s
possible for it to do so without the service having an organisation.

Putting the check here will mean that the user at least gets a `403`
here, rather than a `500`.