Stop go live if organisation hasn’t signed the agreement
We already stop platform admin users from making services live withou
an organisation.

Sometimes they create and organisation, assign the service to it, then
make it live.

This inadvertantly skips the step where the service needs to sign the
agreement on behalf of their organisation.

This commit puts a restriction in place to stop that from being
possible.

Note we still have the concept of a service-specific agreement (defined
as `agreement_signed=None`) which this continues to support. Argument
for another day whether we should remove that concept entirely.