Only show the specific error to platform admins
Only platform admin users should be reaching this page. But the error
we are giving is specific to them (only they can link services to
organisations). So just to future proof this and make sure we don’t
accidentally show this page to a normal user, let’s add another guard
here.