Replace hardcoded 7 day retention statement with service specific
The API integration page was hardcoded to 7 days, which may not
be correct as services can have retentions between 3 and 90
days.

If the service has consistent retention, ie email, letter and sms
have the same retention periods, it is easy to update this number
to make it accurate.

If the service does not have consistent retention, which is rare,
but possible (only platform admins can put a service in this state)
then I've opted for a quick solution of just removing the line
from the page. Happy for this to be adapated as designers see
fit (we could for example show the retention per channel again
if we wanted to, or instead refer generically to the retention
settings for their service).

I've tested this with both live and trial mode services (trial mode
services don't have any rows in the service_data_retention table).