dont run accessibility statement tests if git errors
we check the accessibility statement for local changes by using a
subprocess to run `git diff` against the origin/main branch.

We check the status code of that git diff to determine if there's been
any changes.

However, if we run the tests on a copy of the admin source code that
doesn't have an active git checkout, then the command will error and
will return statuscode 1.

So, to determine if there were changes to the accessibility statement,
we should look at stderr to make sure it didn't output any error
messages