remove .git from .dockerignore
our unit tests (specifically test_accessibility_statement.py) expect to
be run in a folder with a local git checkout, so that it can compare the
local changes to main to make sure that any changes to the accessibility
statement have been made correctly.

however, in the test image we were previously excluding the .git folder
via the dockerignore, so the test would error with an unhelpful message.

as the test image calls `copy . .` it will now pull in the .git folder
so will be able to diff succesfully.

note: the production image only copies across app, application.py,
entrypoint.sh and gunicorn_config.py so our prod tasks won't have the
git folder floating about