remove erroring autospecs
in python 3.11 or before, the autospec arg now errors if we're patching
an already mocked function (as autospec doesn't know what the original
function signature is).

a couple of these I was just able to replace with the original fixture,
as it's patched by mock_login. some though i thought it simpler just to
drop the autospec