Reduce cache TTL to 1 hour
Default is 7 days.

This would mean someone would have to remember to clear the cache on the
day the rates changes.

Rates only change at the start of a day so by making the TTL 1 hour they
will only be incorrect for a maximum of 1 hour from midnight until 1am.
This is a big improvement on our current manual process of updating the
hard coded values with a deploy.

There’s also precedent for using a value of 1 hour on high traffic pages
for the count of live services and organisations:
https://github.com/alphagov/notifications-admin/blob/da3798c02f942c32521ebf24fd10364423727384/app/notify_client/status_api_client.py#L8