Fix API unknown errors
`requests.Response` defines `__bool__` as `self.ok`, which returns True
if the status code is < 400.

Therefore, all 4xx and 5xx responses are considered `False`. Which means
that when we try to log errors, and check `if error.response`, it always
evaluates to `False`, and so we never log the response URL - just
`unknown`.