Use regular method for example
The methods for getting a preview of a normal template and getting a
preview of the special example template are very similar.

This commit moves the view which renders the example template to using
the method which renders the normal template.

It then deletes the special method used for rendering the example
template.