Only call template preview for letter templates
If someone hacks the URL to try to get a preview of an email template
then they will get a 500 back from template preview.

Lets catch this early and return a 404 instead because we have no way
of showing a PNG or PDF for an email or text message template.