Fix letter branding preview when no service present
Because we serve letter branding images without cookies there’s no
`current_service` available in the requet context.

This was causing an error when trying to preview a letter with no
branding, because the fallback is to look at the branding for the
current service.

This commit adds a conditional for this case, and makes the test more
end-to-end to ensure this edge case is accounted for.