Use the internal API host name, rather than the external
For the admin to talk to the API in PaaS, we want it to always use
the PaaS route (the internal one). This keeps things consistent
with ECS where all inter app requests stay in ECS (so we
want all inter app requests in the PaaS to stay in PaaS).

For the moment, this change won't actually do anything because
in the credentials repo the API_HOST_NAME_INTERNAL is still
set to the external address, but we will shortly change the value
to be the PaaS internal address.

This PR Relies on the creation of the internal secret in
alphagov/notifications-credentials#417