Show fullscreentable row headers
We currently hide them but this creates a split
between the visual interface (which shows them)
and the semantic interface accessed by screen
readers.

This:
- stops hiding them
- removes the margin-left we added to allow for
  the space left over by hiding them

Notes on changes to js tests:

This makes the tests match the new behaviour but
also changes the value used for the main column
width. This is now changed to 712px, which is what
it measures on a screen >= 960px, the size most of
our users have. I made this change to help
comparisons with the component in-browser.

Notes on changes to HTML:

This includes changing the first cells for each
row from `td`s to `th`s. This change also requires
us adding `scope=row` but I'm splitting this out
into a separate commit as it's to a macro which is
used by other code.