Fix layout issue making index columns out of sync
The algorithm used by browsers to layout tables
renders the same content in different ways for a
table that is positioned absolutely vs one with
static (default) position.

We position the 'fake' table we use to show fixed
row headers differently to the original. Depending
on the content, this can mean they are laid out
differently. This tends to occur when there are
only a few columns of data and the table has the
desktop screen size to fill.

This adds some JS to detect when this is happening
and fix the width of that column in both tables to
stop it.