Clear the active go-live request when organisation is changed
If we change a service’s organisation it’s usually because we need the
user to accept the data sharing and financial agreement.

After they have done this we ask them to re-submit the request.

If in the future we start blocking duplicate go live requests then we
need to make sure that people who do genuinely need to submit another
request aren’t blocked.