Give go-live Zendesk tickets different subjects
At the moment, all go live requests have a subject of "Request to
go live - service name" and we need to check the ticket content to see
if a request will be self-approved.

This changes the subject for requests that can be self approved to make
them easier to spot: ""Self approve go live request - service name".