Hide the overlapped row header cells
We need to hide the row header cells on the
scrollable table, beneath the fixed row header
cells for some browsers on OSX and iOS.

This is to tidy up the 'bouncy' state you get when
you hit the end of the scrolling because it
overshoots the end of the scroll area and so
reveals those cells. We can't do that with
`visibility: hidden` because it also hides them
from screen readers so have to use a hack where we
set their text and background to the same colour
as the page background and hide the borders.